improvement(hdrh): move 'CSHDRFileLogger' to a common place
We are going to add support of hdr histogram files for other
benchmarking tools too.
So, move the 'CSHDRFileLogger' class to a common place
and rename it to the 'HDRHistogramFileLogger'.

(cherry picked from commit 67eceadb9e4c2b8ad212e3e970f885e97d187a0e)
